3rd person camera will chase target moving forward direction automatically

+++ StenGine/include/Utility/Timer.h
@@ -9,7 +9,7 @@ namespace StenGine
 
 class Timer {
 public:
-	using Seconds = double;
+	using Seconds = float;
 
 	Timer();
 	~Timer();
